requirements.txt
actdiag==0.5.4
ActionCableZwei==0.1.7.2
awscli==1.16.52
beautifulsoup4==4.6.3
blockdiag==1.5.4
boto3==1.9.77
botocore==1.12.77
certifi==2018.11.29
chardet==3.0.4
click==6.7
colorama==0.3.9
docutils==0.14
elasticsearch==6.3.1
elasticsearch-curator==5.6.0
funcparserlib==0.3.6
get==2018.11.19
idna==2.7
jmespath==0.9.3
Pillow==5.3.0
post==2018.11.20
public==2018.11.20
pyasn1==0.4.4
python-dateutil==2.7.5
PyYAML==3.13
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
requests-aws4auth==0.9
rsa==3.4.2
s3transfer==0.1.13
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
voluptuous==0.11.5
webcolors==1.8.1
websocket-client==0.54.0
